FixCarrito.FixAprobaciones

+++ inka-productos/src/main/java/pe/cibertec/inkaproductos/repositories/MovimientoRepository.java
@@ -1,8 +1,7 @@
 package pe.cibertec.inkaproductos.repositories;
 
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.cibertec.inkaproductos.models.Movimiento;
 
-public interface MovimientoRepository extends JpaRepository<Movimiento, Long> {
+public interface MovimientoRepository extends JpaRepository<Movimiento, Integer> {
 }
